UserModelConfig with @property decorator
